X11 - fix segfauls

+++ xbmc/windowing/X11/WinSystemX11GLContext.cpp
@@ -211,13 +211,15 @@ bool CWinSystemX11GLContext::SetFullScreen(bool fullScreen, RESOLUTION_INFO& res
 
 bool CWinSystemX11GLContext::DestroyWindowSystem()
 {
-  m_pGLContext->Destroy();
+  if (m_pGLContext)
+    m_pGLContext->Destroy();
   return CWinSystemX11::DestroyWindowSystem();
 }
 
 bool CWinSystemX11GLContext::DestroyWindow()
 {
-  m_pGLContext->Detach();
+  if (m_pGLContext)
+    m_pGLContext->Detach();
   return CWinSystemX11::DestroyWindow();
 }
 
@@ -296,7 +298,7 @@ bool CWinSystemX11GLContext::RefreshGLContext(bool force)
           return true;
       }
     }
-    else if (gli != "EGL")
+    else if (gli == "EGL_PB")
     {
       success = m_pGLContext->CreatePB();
       if (success)
